+++ src/ralph_pricing/views/usages.py
@@ -60,6 +60,14 @@ class Usages(Base):
 
     def get_context_data(self, **kwargs):
         context = super(Usages, self).get_context_data(**kwargs)
+
+        if self.usage_type:
+            by_warehouse = self.usage_type.by_warehouse
+        else:
+            by_warehouse = None
+
+        by_cost = self.usage_type.by_cost if self.usage_type else None
+
         context.update({
             'section': 'usages',
             'sidebar_items': usages_menu(
@@ -67,6 +75,8 @@ class Usages(Base):
                 self.usage_type_name
             ),
             'sidebar_selected': self.usage_type_name,
+            'by_cost': by_cost,
+            'by_warehouse': by_warehouse,
             'formset': self.formset,
         })
         return context