+++ src/ralph_pricing/views/ventures.py
@@ -11,18 +11,38 @@ from django.utils.translation import ugettext_lazy as _
 
 from ralph_pricing.views.reports import Report, currency
 from ralph_pricing.models import UsageType, ExtraCostType, Venture
-from ralph.business.models import Venture as ralph_venture
 from ralph_pricing.forms import DateRangeForm
 
 
 class AllVentures(Report):
+    '''
+        Reports for all ventures
+    '''
     template_name = 'ralph_pricing/ventures_all.html'
     Form = DateRangeForm
     section = 'all-ventures'
     report_name = _('All Ventures Report')
 
     @staticmethod
-    def get_data(start, end, show_in_ralph=False, **kwargs):
+    def get_data(
+        warehouse,
+        start,
+        end,
+        show_in_ralph=False,
+        forecast=False,
+        **kwargs
+    ):
+        '''
+            Generate raport for all ventures
+
+            :param integer warehouse: Id warehouse for which is generate report
+            :param datetime start: Start of the time interval
+            :param datetime end: End of the time interval
+            :param boolean show_in_ralph: if true, show only active ventures
+            :param boolean forecast: if true, generate forecast raport
+            :returns list: List of lists with report data and percent progress
+            :rtype list:
+        '''
         # 'show_in_ralph' == 'show only active' checkbox in gui
         ventures = Venture.objects.order_by('name')
         total_count = ventures.count() + 1  # additional step for post-process
@@ -30,14 +50,9 @@ class AllVentures(Report):
         totals = {}
         values = []
         for i, venture in enumerate(ventures):
-            try:
-                show_venture = ralph_venture.objects.get(
-                    id=venture.venture_id,
-                ).show_in_ralph
-            except ralph_venture.DoesNotExist:
-                show_venture = False
-            if show_in_ralph and not show_venture:
+            if show_in_ralph and not venture.is_active:
                 continue
+
             values_row = {}
             values.append(values_row)
             count, price, cost = venture.get_assets_count_price_cost(
@@ -49,7 +64,7 @@ class AllVentures(Report):
             row = [
                 venture.venture_id,
                 path,
-                show_venture,
+                venture.is_active,
                 venture.department,
                 venture.business_segment,
                 venture.profit_center,
@@ -63,6 +78,8 @@ class AllVentures(Report):
                     start,
                     end,
                     usage_type,
+                    warehouse.id if usage_type.by_warehouse else None,
+                    forecast=forecast,
                 )
                 row.append(count)
                 column += 1
@@ -124,12 +141,33 @@ class AllVentures(Report):
 
 
 class TopVentures(AllVentures):
+    '''
+        Reports for top ventures
+    '''
     template_name = 'ralph_pricing/ventures_top.html'
     section = 'top-ventures'
     report_name = _('Top Ventures Report')
 
     @staticmethod
-    def get_data(start, end, show_in_ralph=False, **kwargs):
+    def get_data(
+        warehouse,
+        start,
+        end,
+        show_in_ralph=False,
+        forecast=False,
+        **kwargs
+    ):
+        '''
+            Generate raport for top ventures
+
+            :param integer warehouse: Id warehouse for which is generate report
+            :param datetime start: Start of the time interval
+            :param datetime end: End of the time interval
+            :param boolean show_in_ralph: if true, show only active ventures
+            :param boolean forecast: if true, generate forecast raport
+            :returns list: List of lists with report data and percent progress
+            :rtype list:
+        '''
         # 'show_in_ralph' == 'show only active' checkbox in gui
         ventures = Venture.objects.root_nodes().order_by('name')
         total_count = ventures.count() + 1  # additional step for post-process
@@ -137,13 +175,7 @@ class TopVentures(AllVentures):
         totals = {}
         values = []
         for i, venture in enumerate(ventures):
-            try:
-                show_venture = ralph_venture.objects.get(
-                    id=venture.venture_id
-                ).show_in_ralph
-            except ralph_venture.DoesNotExist:
-                show_venture = False
-            if show_in_ralph and not show_venture:
+            if show_in_ralph and not venture.is_active:
                 continue
             values_row = {}
             values.append(values_row)
@@ -155,7 +187,7 @@ class TopVentures(AllVentures):
             row = [
                 venture.venture_id,
                 venture.name,
-                show_venture,
+                venture.is_active,
                 venture.department,
                 venture.business_segment,
                 venture.profit_center,
@@ -169,6 +201,8 @@ class TopVentures(AllVentures):
                     start,
                     end,
                     usage_type,
+                    warehouse.id if usage_type.by_warehouse else None,
+                    forecast=forecast,
                     descendants=True,
                 )
                 row.append(count)